add word hint to the guessers (every 10 second radom letter from the word will show up)

+++ server/src/main/java/com/example/drawandguess/listener/WebSocketEventListener.java
@@ -25,7 +25,9 @@ public class WebSocketEventListener {
         StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
         String sessionId = accessor.getSessionId();
         Participant participant = participantService.findParticipantBySessionId(sessionId);
-        roomService.removeParticipantFromAllRooms(participant);
-        participantService.removeParticipant(sessionId);
+        if (participant != null) {
+            roomService.removeParticipantFromAllRooms(participant);
+            participantService.removeParticipant(sessionId);
+        }
     }
 }
